Add chown to node_modules copy
See linked ticket for more information; npm packages appear to have all
sorts of weird user/group combinations, which will occasionally break
containers. Given that we're executing as a single user, and given that
we used to provision these with root/root, it should be safe to chown
all of these files on copy.

This should fix issues pulling the container with Podman 3.4.4 on
Ubuntu 22.04.

Resolves: #1491

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>