Update consul-associate-003.mdx
Added caveat to Requirements section because the requirements to take an exam will change as of May 1st. If a customer were to follow this today, and documentation is currently instructing to PSI's requirements, they won't have an accurate understanding of what's needed for Certiverse.

While PSI's systems are more restrictive and someone who preps to PSI's standards are *probably* fine in Certiverse's more lax systems, a change in day-of exam experience vs expectation can be jarring.