Move private repo logic from server.ts to loader
Since logic was implemented in the loader to hide the "Edit this page on
GitHub" link for PTFE, it makes sense to move the other (similar) logic
there too to keep everything together.

Additionally, this also allows the expression to be cleaned up to use
`this.opts.product` and `Array.includes()` to detect private repos.