revert changes to sentinel.json product data
Rationale for reverting th basePaths change: we only ever use product data basePaths for redirects from the previous dot-io sites. This property should likely be renamed, or perhaps removed entirely, as redirects could probably be written in a more static way now that we've completed our migration to Dev Dot.

Rationale for reverting the rootDocsPath change: up until the point we fully remove intro content, including the page files, we need to maintain this data, otherwise the build will fail. Recommendation is to remove intro content in a separate PR.