Update playwright version to fix e2e runner
An outdated version of playwright had dependencies that were no longer
present in the package repositories for ubuntu@latest (noble).
Downgrading to ubuntu 22.04 temporarily to see if playwright would
install on 22.04 confirmed this.

The solution ended up being to upgrade playwright so that we can
continue to use ubuntu@latest