Port the eslint config from UDR over to devdot
Porting the linting setup from UDR was deemed to be easier and faster than
attempting to modify `hashicorp-next`, republish and then try to integrate the
new package version.

The eslint config from UDR was written as an esmodule in order to use things
like import - so a straight copy and paste wasn't possible. But with a little
massaging (i.e: swapping out `import` for require) it seems to work in a
rudimentary sense.

Additionally, some tweaks had to be made to package.json to uninstall
`hashicorp-next` and remove references to it in pre-commit hooks(in
order to be able to commit)