Port the eslint config from UDR and next-hashicorp (#2704)
* Port the eslint config from UDR and next-hashicorp

Porting the linting setup from UDR was deemed to be easier and faster than
attempting to modify `hashicorp-next`, republish and then try to integrate the
new package version.

The eslint config from UDR was written as an esmodule in order to use things
like import - so a straight copy and paste wasn't possible. But with a little
massaging (i.e: swapping out `import` for require) it seems to work in a
rudimentary sense.

Additionally, some tweaks had to be made to package.json to uninstall
`hashicorp-next` and remove references to it in pre-commit hooks(in
order to be able to commit)

Finally the old rules from next-hashicorp were ported over in attempt to
resolve lint issues and bring the lint rules in line with the current
code style in devdot

* Remove content directory from eslint ignore

Devdot does not have a content directory, so there's no need to have it
in the ignore list

* Change vitest config file extension

The vitest config file has an extension of `.mts`, not `.mjs`