Ignore .extracted
This directory contains compiled/transpiled code that (likely) contains
multiple rule violations. The most notable example being
`public/ie-warning.js` which contains a declared (but unused in that
file) funciton. This entire file is intended to be linked from a web
page using <script> tags, but it's existence triggers a
`no-unused-expressions` eslint error.

The solution therefore is to ignore the entire .extracted directory
since the entire directory is generated anyway.