Resolve remaining lint issues (#2716)
* Port the eslint config from UDR and next-hashicorp (#2704)

* Port the eslint config from UDR and next-hashicorp

Porting the linting setup from UDR was deemed to be easier and faster than
attempting to modify `hashicorp-next`, republish and then try to integrate the
new package version.

The eslint config from UDR was written as an esmodule in order to use things
like import - so a straight copy and paste wasn't possible. But with a little
massaging (i.e: swapping out `import` for require) it seems to work in a
rudimentary sense.

Additionally, some tweaks had to be made to package.json to uninstall
`hashicorp-next` and remove references to it in pre-commit hooks(in
order to be able to commit)

Finally the old rules from next-hashicorp were ported over in attempt to
resolve lint issues and bring the lint rules in line with the current
code style in devdot

* Remove content directory from eslint ignore

Devdot does not have a content directory, so there's no need to have it
in the ignore list

* Change vitest config file extension

The vitest config file has an extension of `.mts`, not `.mjs`

* Fix no-unused-vars (#2705)

@typescript-eslint/no-unused-vars

Fix all instances of the above lint rule

* @typescript-eslint/no-explicit-any (#2706)

Fix linting for the above eslint rule

* @typescript-eslint/no-require-imports (#2707)

* Fix no-require-imports

Looking at the code for www it seems as though `require()` is used there
for image imports. Under closer inspection, it would seem as though this
rule was mistakenly only applied to `.js` files rather than all source
files.

Changing that resolved the lint errors for this rule

* Fix indentation for eslint config file

* @typescript-eslint/no-unused-expressions (#2709)

* @typescript-eslint/no-unused-expressions

* Ignore .extracted

This directory contains compiled/transpiled code that (likely) contains
multiple rule violations. The most notable example being
`public/ie-warning.js` which contains a declared (but unused in that
file) funciton. This entire file is intended to be linked from a web
page using <script> tags, but it's existence triggers a
`no-unused-expressions` eslint error.

The solution therefore is to ignore the entire .extracted directory
since the entire directory is generated anyway.

* Add react prop types (#2711)

* react/prop-types

Add prop type declarations to react components to resolve
react/prop-types lint error. This also required the installation of the
prop-types package since React no longer provides this

* Fix vercel build

Mistakenly used `string` instead of `Proptypes.string` causing the
vercel build to fail due to `string` not being defined.

* no-array-index-key (#2714)

* no-useless-escape

* valid-typeof

* react-hooks/exhaustive-deps

- Installs the eslint-plugin-react hooks eslint plugin in order to provide
  guidance around the correct usage of react hooks

- Resolve eslint eshaustive-deps issues
  Effect hooks should be provided with an array of the hook
  dependencies(as-per recommendations from the react team) in order to
  avoid unexpected behaviour if/when those dependencies change.

- Alias portalRef.current
  Alias portalRef.current outside cleanup function to avoid issues
  potentially caused by the ref being removed by the time the clean-up
  function is run

* @next/next/no-img-element

* react-hooks/rules-of-hooks

* no-use-before-define

The test file attempts to call `vi.stubEnv` before `vi` has been
imported. Since the environment var being stubbed is stubbed in the test
where it is being used (and it's removal doesn't impact the tests) it
looks as though it can safely be removed

* import/no-anonymous-default-export

* @typescript-eslint/no-empty-object-type

* prefer-const

* Use `<img>` element to display user avatar

Usage of the `<Image />` component from next caused issues with loading
images from unapproved remote domains (gravatar for instance) when
signing in with GitHub.

For now the safest option is to revert that change, use the <img> element
instead and perhaps revisit this at a later date...

This commit partially reverts e6820b30a69dac062cce3d94cd049cb1d4a464fe

* no-prototype-builtins (#2710)

* no-prototype-builtins

Fix eslint rule violations concerning the usage of foo.hasOwnProperty('bar')

* Revert "no-prototype-builtins"

Following discussion on 3/17 - it was decided to revert this commit and
simply turn this rule off to be more inline with our current coding
practices

* Turn off no-prototype-builtins

* Empty commit to re-trigger CI