fix: un-revert #2708 and fix package-lock file (#2728)
* Port the eslint config from UDR and next-hashicorp (#2704)

* Port the eslint config from UDR and next-hashicorp

Porting the linting setup from UDR was deemed to be easier and faster than
attempting to modify `hashicorp-next`, republish and then try to integrate the
new package version.

The eslint config from UDR was written as an esmodule in order to use things
like import - so a straight copy and paste wasn't possible. But with a little
massaging (i.e: swapping out `import` for require) it seems to work in a
rudimentary sense.

Additionally, some tweaks had to be made to package.json to uninstall
`hashicorp-next` and remove references to it in pre-commit hooks(in
order to be able to commit)

Finally the old rules from next-hashicorp were ported over in attempt to
resolve lint issues and bring the lint rules in line with the current
code style in devdot

* Remove content directory from eslint ignore

Devdot does not have a content directory, so there's no need to have it
in the ignore list

* Change vitest config file extension

The vitest config file has an extension of `.mts`, not `.mjs`

* Fix no-unused-vars (#2705)

@typescript-eslint/no-unused-vars

Fix all instances of the above lint rule

* @typescript-eslint/no-explicit-any (#2706)

Fix linting for the above eslint rule

* @typescript-eslint/no-require-imports (#2707)

* Fix no-require-imports

Looking at the code for www it seems as though `require()` is used there
for image imports. Under closer inspection, it would seem as though this
rule was mistakenly only applied to `.js` files rather than all source
files.

Changing that resolved the lint errors for this rule

* Fix indentation for eslint config file

* @typescript-eslint/no-unused-expressions (#2709)

* @typescript-eslint/no-unused-expressions

* Ignore .extracted

This directory contains compiled/transpiled code that (likely) contains
multiple rule violations. The most notable example being
`public/ie-warning.js` which contains a declared (but unused in that
file) funciton. This entire file is intended to be linked from a web
page using <script> tags, but it's existence triggers a
`no-unused-expressions` eslint error.

The solution therefore is to ignore the entire .extracted directory
since the entire directory is generated anyway.

* Add react prop types (#2711)

* react/prop-types

Add prop type declarations to react components to resolve
react/prop-types lint error. This also required the installation of the
prop-types package since React no longer provides this

* Fix vercel build

Mistakenly used `string` instead of `Proptypes.string` causing the
vercel build to fail due to `string` not being defined.

* no-array-index-key (#2714)

* no-prototype-builtins (#2710)

* no-prototype-builtins

Fix eslint rule violations concerning the usage of foo.hasOwnProperty('bar')

* Revert "no-prototype-builtins"

Following discussion on 3/17 - it was decided to revert this commit and
simply turn this rule off to be more inline with our current coding
practices

* Turn off no-prototype-builtins

* copy everything from #2716 except package-lock and regen that file instead

* Disable react/prop-types rule completely

After discussion with @heatlikeheatwave we couldn't figure out where this rule
being applied only to JSX actually came from in the first place. The best we
could do is trace it to
https://www.npmjs.com/package/@hashicorp/next-hashicorp?activeTab=code where it
looks like it's disabled completely.

However, between `next-hashicorp` and the UDR lint rules, we couldn't find
anywhere where it's only applied to JSX. Therefore, most of those
changes have been reverted and the rule turned off for all source files

This commit partially reverts e5be7eb2e22bbf1faa2729b619cadbe2ed7e7c89

* Add .vercel to eslint ignore list

* fix copy paste oops

* remove unnecessary ref comment

---------

Co-authored-by: Robert Main <50675045+rmainwork@users.noreply.github.com>