Fetch versions for current IA as well as previous
Retrieves a list of valid versions of the current document using the
current path that we have as well as the previous path. This is to fix
an issue where versions at the new location were missing.

In this instance from v1.18.x to v1.19.x the documentation moved from
folderA to folderB and only the versions present in folderA were being
returned