Send user to home page for missing documents
Next redirects are written using Vercel's own dialect of regex. This
means that the complex (i.e: /vault/foo/(:path) type) redirects can't
simply be fed to `new RegExp()` as I was hoping and parsing them
programatically is a more challenging endevour than originally thought.

Therefore, the best stopgap fix for now is to simply send the user to
the home page in the event that no versions are available for a given
document