Updates Node.js version to v22 (#2821)
* Updates Node.js version to v22

Sets the Node.js version to v22 across all relevant workflow files and the Dockerfile.
This ensures that the project is using the latest LTS version of Node.js for improved performance, security, and compatibility.
Also, updates the engines field in `package.json` to reflect the new Node.js version.

* Improves image optimization and API handling

- Adds optional chaining to prevent errors when tutorials are undefined.
- Removes dev dependencies.
- Updates image optimization config to correctly use env variables.

This change fixes an issue where the tutorials property in the collection data was sometimes undefined, leading to an error in the UI. The optional chaining ensures that the map function is only called if tutorials is an array.
Removes redundant dev dependencies from the project, cleaning up the codebase and potentially reducing build times.

* Adds missing source-map dependency

Ensures escodegen includes its source-map dependency. This resolves potential issues related to missing or outdated source map functionality within the escodegen library.

* Updates base image to Bookworm

Updates the base image for both the deps and builder stages
from `buster` to `bookworm`.

This aligns the Dockerfile with the latest supported
Node.js image and incorporates security updates and
improvements available in the newer Bookworm distribution.