Removes "Other Sandboxes" section (#2838)
* Temporarily removes "Other Sandboxes" section

Removes the "Other Sandboxes" section from the sandbox dropdown.

This change is part of a redesign effort and the section will be re-evaluated and potentially re-introduced later. The related test is also temporarily removed.

* [COMPLIANCE] Add Copyright and License Headers (#2733)

Co-authored-by: hashicorp-copywrite[bot] <110428419+hashicorp-copywrite[bot]@users.noreply.github.com>

* Removes other sandboxes from dropdown

Removes the "Other Sandboxes" section from the sandbox dropdown in preparation for a redesign.
This change removes the related UI elements and associated logic.

---------

Co-authored-by: hashicorp-copywrite[bot] <110428419+hashicorp-copywrite[bot]@users.noreply.github.com>