Et/fix sandbox launch navigation (#2889)
* Refactors Instruqt lab integration for tutorials

Moves Instruqt context provider to the tutorial view.
This change ensures that the lab ID is correctly passed and managed,
improving the reliability of interactive tutorials.

Removes unused code and simplifies logic related to fallback lab IDs.

* Simplifies Instruqt lab opening logic

Removes unnecessary checks and state management related to lab configuration errors and lab existence when opening an Instruqt lab. This simplifies the logic and improves the overall efficiency of the lab opening process. The state management was causing unnecessary complexity.

* Adds className prop to CardsGridList component

Enables passing custom CSS classes to the CardsGridList component.

This change allows for greater flexibility in styling the CardsGridList, as demonstrated by its use in the SandboxView to define a two-column layout on tablet devices and above.

Also, it disables Instruqt lab iframe when not in interactive mode.

* Restores lab state from local storage

Attempts to restore the Instruqt lab ID and active state from local storage on page load.

Adds error handling and tracking for storage access failures, including attempts to clear corrupted storage. This ensures a smoother user experience in cases where the stored state is invalid.

* Refactors Instruqt lab integration for tutorials

Moves Instruqt context provider to the tutorial view.
This change ensures that the lab ID is correctly passed and managed,
improving the reliability of interactive tutorials.

Removes unused code and simplifies logic related to fallback lab IDs.

* Simplifies Instruqt lab opening logic

Removes unnecessary checks and state management related to lab configuration errors and lab existence when opening an Instruqt lab. This simplifies the logic and improves the overall efficiency of the lab opening process. The state management was causing unnecessary complexity.

* Adds className prop to CardsGridList component

Enables passing custom CSS classes to the CardsGridList component.

This change allows for greater flexibility in styling the CardsGridList, as demonstrated by its use in the SandboxView to define a two-column layout on tablet devices and above.

Also, it disables Instruqt lab iframe when not in interactive mode.

* Restores lab state from local storage

Attempts to restore the Instruqt lab ID and active state from local storage on page load.

Adds error handling and tracking for storage access failures, including attempts to clear corrupted storage. This ensures a smoother user experience in cases where the stored state is invalid.

* Removes unused React import

Removes an unnecessary import statement for React in the tutorial view sidebar content component.